+++ CHANGELOG.md
@@ -20,6 +20,7 @@ There are following change types:
 
 - Add new plugin integrations:
     - 'folke/snacks.nvim'
+    - 'saghen/blink.cmp'
 
 ## mini.clue